+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat: various output handler output id and project name
 - feat: Respect `CODEX_HOME` environment variable when calculating session path
 - feat: New token handler `notify` that sends native notifications on windows
+- Add bundle script that creates an application pyz bundle
 
 ### Changed
 - refactor: rename core.`SessionState` to `SessionContext`